msxbatoken: Remove INI file
It was adding a non-trivial amount of code and was not pulling its weight,
particular since the only place the INI file was read from was from the
directory in which the program resided, making it not-so-useful in any
situation in which the binary is shared amongst different users or just
different projects.

It's easy enough to call this as part of a script or a build system if the
command line parameters need to be recorded in a file.