msxbatoken: Bring output filename handling into Args class
This also simplifies a bunch of code, and restores the functionality of
being able to provide one filename and have output go to the same path with
the extension replaced by `.bas`. (I don't know where I broke that, or if
it was always broken, but it's not worth investigating.)

This does now change the failure message behaviour, in that we print a
traceback instead of just an error message when we refuse to overwrite the
input file. We'll leave this as is at the moment becuase the whole error
message system (which is mixed in with the logging system, including the
logging system exiting on certain log levels) needs to be completely
reworked anyway.