msxbatoken: Better fatal error handling, including proper return code
Use a separate function for fatal errors, making it more clear that
fatal errors exit the program. (Previously this was done by logging a
message at level 1 and adding a `# Exit` comment to the line to
indicate that the show_log exits for that level only.) The function
takes a list of message parts, so we no longer need to join them at
every call site.

Send the error message to stderr instead of stdout, and exit with exit
code 1, indicating an error occurred. Previously we exited with code 0,
indicating success, even though we had failed.