msxbatoken: Clean up generation of the head and tail of the list file
The formatting of the hex and decimal display of the start/end/size is a
bit dodgy: `&h014F > 335` looks like it's saying the left is greater than
the right. I think that `&h014F (335)` would be more clear, but I've not
changed it because I don't know if the tools that read .mlb files read
those lines and expect the current formatting.